requirement.txt
astroid==2.2.5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
django==2.0.4
django-bootstrap3==11.1.0
django-braces==1.13.0
django-heroku==0.3.1
django-jet==1.0.8
django-misaka==0.2.1
djangorestframework==3.10.0
google-api-python-client==1.4.1
houdini.py==0.1.0
httplib2==0.13.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
misaka==2.1.1
numpy==1.16.0
oauth2client==2.2.0
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
Pygments==2.4.2
PyLin==0.2.7
pylint==2.3.1
pyserial==3.4
pytz==2018.4
requests==2.22.0
rsa==4.0
six==1.12.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.6.1
whitenoise==4.1.3
wrapt==1.11.2
